chore: move around a few more components and mock data
